add security layer to backend routes including POST, DELETE and PUT routes and update front-end components

+++ server/api/carts.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const {
-	models: { Book, Cart, CartItem, User }
+	models: { Cart, CartItem, User }
 } = require('../db');
 
 // GET /api/carts
@@ -17,55 +17,73 @@ router.get('/pending', async (req, res, next) => {
 	}
 });
 
-// POST /api/carts/add/books/:bookId
+// POST /api/carts/add/books
 // Add a book to a cart
-router.post('/add/books/:bookId', async (req, res, next) => {
+router.post('/add/books', async (req, res, next) => {
 	const token = req.headers.authorization;
 	const user = await User.findByToken(token);
 	if (user) {
 		try {
 			const cart = await Cart.findPendingCartForUser(user.id);
 
-			await CartItem.create({
-				quantity: 1,
-				cart_id: cart.id,
-				book_id: req.params.bookId
+			await CartItem.findOrCreate({
+				where: {
+					cart_id: cart.id,
+					book_id: req.body.id
+				},
+				defaults: {
+					cart_id: cart.id,
+					book_id: req.body.id
+				}
 			});
 			res.sendStatus(201);
 		} catch (error) {
 			next(error);
 		}
-}});
-
-// DELETE /api/carts/:cartId/books/:bookId
-// remove book(s) from a cart
-router.delete('/:cartId/books/:bookId', async (req, res, next) => {
-	try {
-		await CartItem.destroy({
-			where: {
-				cart_id: req.params.cartId,
-				book_id: req.params.bookId
-			}
-		});
-		res.status(204).send('book removed from your cart');
-	} catch (error) {
-		next(error);
 	}
 });
 
-// PUT /api/carts/:cartId/books/:bookId
+// DELETE /api/carts/remove/books
+// Remove book(s) from a cart
+router.delete('/remove/books', async (req, res, next) => {
+	const token = req.headers.authorization;
+	const user = await User.findByToken(token);
+	if (user) {
+		try {
+			const cart = await Cart.findPendingCartForUser(user.id);
+			await CartItem.destroy({
+				where: {
+					cart_id: cart.id,
+					book_id: req.body.id
+				}
+			});
+			res.sendStatus(204);
+		} catch (error) {
+			next(error);
+		}
+	}
+});
+
+// PUT /api/carts/books/quantity
 // Edit quantity in a cart
-router.put('/:cartId/books/:bookId', async (req, res, next) => {
-	try {
-		const cart_item = await CartItem.findOne({
-			where: {
-				cart_id: req.params.cartId,
-				book_id: req.params.bookId
+router.put('/books/quantity', async (req, res, next) => {
+	const token = req.headers.authorization;
+	const user = await User.findByToken(token);
+	if (user) {
+		try {
+			const cart = await Cart.findPendingCartForUser(user.id);
+			if (cart) {
+				const cartItem = await CartItem.findOne({
+					where: {
+						cart_id: cart.id,
+						book_id: req.body.id
+					}
+				});
+				res.json(await cartItem.update(req.body));
 			}
-		})
-		res.json(await cart_item.update(req.body));
-	} catch (error) {
-		next(error);
+		} catch (error) {
+			next(error);
+		}
 	}
 });
